+++ src/main/java/org/bytedeco/javacpp/annotation/ArrayAllocator.java
@@ -1,5 +1,6 @@
 package org.bytedeco.javacpp.annotation;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -23,6 +24,6 @@ import org.bytedeco.javacpp.tools.Generator;
  *
  * @author Samuel Audet
  */
-@Retention(RetentionPolicy.RUNTIME)
+@Documented @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD})
 public @interface ArrayAllocator { }