+++ src/main/java/org/bytedeco/javacpp/annotation/Index.java
@@ -1,5 +1,6 @@
 package org.bytedeco.javacpp.annotation;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -15,7 +16,7 @@ import org.bytedeco.javacpp.tools.Generator;
  *
  * @author Samuel Audet
  */
-@Retention(RetentionPolicy.RUNTIME)
+@Documented @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD})
 public @interface Index {
     /** The number of indices spread over the parameters, for multidimensional access. */